CHANGE: Move module contents to two.py, destined to be imported only if python < 3 is running.
