+++ unicodecsv/test_deps.py
@@ -1,8 +1,11 @@
 import sys
+old = sys.version_info[0] < 3
 
-if sys.version_info[0] < 3:
+if old:
     import unittest2 as unittest
     from StringIO import StringIO
+    from string import letters
 else:
     import unittest
     from io import StringIO
+    from string import ascii_letters as letters
